requirements.txt
# Core dependencies
python-dotenv>=1.0.1
tqdm>=4.66.2

# LangChain ecosystem
langchain>=0.2.0
langchain-core>=1.0.1
langchain-community>=0.2.0
langchain-openai>=0.1.7
langchain-text-splitters>=0.2.0

# Vector DB
chromadb>=0.5.0
# Streamlit app
streamlit>=1.28.0
